Backend/requirements.txt
Flask==2.3.2
google_api_python_client==2.105.0
nltk==3.8.1
numpy==1.26.1
pandas==2.0.2
python-dotenv==1.1.1
tensorflow==2.13.0rc0
tensorflow_macos==2.13.0rc0
